主要是完成“我的”部分 1. 完成个人信息 2. 完成账号信息 3. 完成教育信息 4. 档案文件一级目录

+++ VisionaryEdu/VisionaryEdu/Student_Parent/GradeTendingViewCtrl.m
@@ -95,7 +95,7 @@
     self.lineChart.valueLabelPosition = ValueLabelRight;
     
     self.lineChart.labelForIndex = ^(NSUInteger item) {
-        XLog(@"x轴 x = %lu",item);
+        XLog(@"x轴 x = %lu",(unsigned long)item);
         return [XDateLabel[item] substringFromIndex:5];
     };
